requirements.txt
pysnowball
requests
flask
python-dotenv
pandas
numpy
matplotlib